backend/requirements.txt
fastapi
uvicorn
openai==1.30.5
python-dotenv==1.0.1
httpx==0.27.0